Warning fixes:         * reload1.c (reload_cse_regs): Cast first arg of `bzero' to char *.         * sdbout.c: Include output.h and toplev.h.         (PUT_SDB_INT_VAL): Use HOST_WIDE_INT_PRINT_DEV to print argument         `a'.  Cast `a' to HOST_WIDE_INT to force it to always be so.         (PUT_SDB_SIZE): Likewise.         * sdbout.h (sdbout_mark_begin_function): Add prototype.         * stmt.c (check_for_full_enumeration_handling): Cast argument of         `warning' to long and use %ld specifier.         * toplev.c (main): Likewise for `fprintf'.         * toplev.h (output_file_directive): Add prototype.         * unroll.c (unroll_loop): Use HOST_WIDE_INT_PRINT_DEC specifier in         call to `fprintf'.         (precondition_loop_p): Likewise.         * varasm.c Include sdbout.h.         (assemble_static_space): Move sometimes-unused variable `rounded'         into the scope in which it is used.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@20651 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/stmt.c
@@ -4524,11 +4524,11 @@ check_for_full_enumeration_handling (type)
 	if (!chain)
 	  {
 	    if (TYPE_NAME (type) == 0)
-	      warning ("case value `%d' not in enumerated type",
-		       TREE_INT_CST_LOW (n->low));
+	      warning ("case value `%ld' not in enumerated type",
+		       (long) TREE_INT_CST_LOW (n->low));
 	    else
-	      warning ("case value `%d' not in enumerated type `%s'",
-		       TREE_INT_CST_LOW (n->low),
+	      warning ("case value `%ld' not in enumerated type `%s'",
+		       (long) TREE_INT_CST_LOW (n->low),
 		       IDENTIFIER_POINTER ((TREE_CODE (TYPE_NAME (type))
 					    == IDENTIFIER_NODE)
 					   ? TYPE_NAME (type)
@@ -4544,11 +4544,11 @@ check_for_full_enumeration_handling (type)
 	    if (!chain)
 	      {
 		if (TYPE_NAME (type) == 0)
-		  warning ("case value `%d' not in enumerated type",
-			   TREE_INT_CST_LOW (n->high));
+		  warning ("case value `%ld' not in enumerated type",
+			   (long) TREE_INT_CST_LOW (n->high));
 		else
-		  warning ("case value `%d' not in enumerated type `%s'",
-			   TREE_INT_CST_LOW (n->high),
+		  warning ("case value `%ld' not in enumerated type `%s'",
+			   (long) TREE_INT_CST_LOW (n->high),
 			   IDENTIFIER_POINTER ((TREE_CODE (TYPE_NAME (type))
 						== IDENTIFIER_NODE)
 					       ? TYPE_NAME (type)
